2-Fix-Profile-Logo
* Fixed positioning of profile logo
* CSS refactor + addition
* New folder for components, for better readability
