+++ eliis-front-end/package.json
@@ -21,6 +21,7 @@
     "mdb-react-ui-kit": "^6.3.0",
     "mdi-react": "^9.3.0",
     "node-json-db": "^2.2.0",
+    "postcss": "^8.4.31",
     "react": "^18.2.0",
     "react-bootstrap": "^2.9.0",
     "react-dom": "^18.2.0",
