--- eliis-front-end/src/components/navbar/Navbar/personal.css
@@ -1,55 +1,0 @@
-.gap-2 {
-    gap: 0.5rem;
-}
-
-.place-items-end {
-    place-items: end;
-}
-
-
-
-.grid-rows-1 {
-    grid-template-rows: repeat(1, minmax(0, 1fr));
-}
-
-@media (min-width: 768px) {
-    .md\:grid-cols-\[2fr_1fr\] {
-      grid-template-columns: 2fr 1fr;
-    }
-}
-
-.font-semibold {
-    font-weight: 600;
-    border-left: 1px solid #ccc;
-  }
-
-.flex-col {
-    flex-direction: column;
-}
-
-
-
-.text-white {
-    --tw-text-opacity: 1;
-    color: rgb(255 255 255 / var(--tw-text-opacity));
-}
-
-.text-2xl {
-    font-size: 1.5rem;
-    line-height: 2rem;
-}
-  
-.col-start-1 {
-    grid-column-start: 1;
-}
-.w-5 {
-    width: 1.25rem;
-}
-
-.text-secondary {
-    background-color: rgb(166 194 100);
-}
-.text-sm {
-    font-size: 0.875rem;
-    line-height: 1.25rem;
-}
